# Enable ImmediateVerification and make it useable