--- a/gap/generic/kernel.gi
+++ b/gap/generic/kernel.gi
@@ -16,34 +16,109 @@
 ##
 #############################################################################
 
-InstallGlobalFunction( FindKernelRandom,
-  function(ri,n)
-    local i,l,rifac,s,x,y;
-    Info(InfoRecog,2,"Creating ",n," random generators for kernel.");
-    l := gensN(ri);
-    rifac := ImageRecogNode(ri);
-    for i in [1..n] do
-        x := RandomElm(ri,"KERNELANDVERIFY",true).el;
+# Helper function for FindKernelRandom and RecogniseGeneric.
+# Generate `n` random kernel elements for `ri` and return a boolean. If
+# `doVerification` is `false`, then add the kernel elements directly to `gensN(ri)`.
+# Otherwise, ask the kernel node to write the random kernel elements as SLPs in
+# the kernel node's nice generators - we call this "verification of the kernel"
+# - and only add a random kernel element to `gensN(ri)`, if it was not possible
+# to write it as an SLP.
+# The return value is `fail`, iff computing an SLP for a random element of the
+# group behind the image node of `ri` failed. This indicates, that something
+# went wrong during recognition of the image.
+# If computing SLPs in the image node worked, then:
+# - if `doVerification` is `true`, return whether the kernel could be verified,
+# - if `doVerification` is `false`, return `true`.
+BindGlobal( "GenerateRandomKernelElementsAndOptionallyVerifyThem",
+  function(ri, n, doVerification)
+    local gens, verificationSuccess, x, s, y, z, i;
+    gens := gensN(ri);
+    verificationSuccess := true;
+    # We generate a random element of the kernel as the quotient of a random
+    # element and the preimage of its image under the homomorphism.
+    for i in [1 .. n] do
+        # Finding kernel generators and immediate verification must use
+        # different random elements! This is ensured by using the same stamp
+        # in both situations.
+        x := RandomElm(ri, "GenerateRandomKernelElementsAndOptionallyVerifyThem", true).el;
         Assert(2, ValidateHomomInput(ri, x));
-        s := SLPforElement(rifac,ImageElm( Homom(ri), x!.el ));
+        s := SLPforElement(ImageRecogNode(ri),
+                                 ImageElm(Homom(ri), x!.el));
         if s = fail then
-            return false;
+            return fail;
         fi;
         y := ResultOfStraightLineProgram(s, ri!.pregensfacwithmem);
-        Add(l,x^-1*y);
-        if InfoLevel(InfoRecog) >= 2 then
-            Print(".\c");
+        z := x^-1*y;
+        if isone(ri)(z) or ForAny(gens, x -> isequal(ri)(x, z)) then
+            continue;
+        fi;
+        if not doVerification or SLPforElement(KernelRecogNode(ri), z!.el) = fail then
+            Add(gens, z);
+            verificationSuccess := not doVerification;
         fi;
     od;
-    if InfoLevel(InfoRecog) >= 2 then
-        Print("\n");
+    return verificationSuccess;
+  end );
+
+InstallGlobalFunction( ImmediateVerification,
+  function(ri)
+    local verified;
+    verified := GenerateRandomKernelElementsAndOptionallyVerifyThem(
+        ri,
+        RECOG_NrElementsInImmediateVerification,
+        true
+    );
+    if verified = fail then
+        ErrorNoReturn("Very bad: image was wrongly recognised ",
+                      "and  we found out too late");
     fi;
-    return true;
+    if verified = true then return true; fi;
+    # Now, verified = false.
+    Info(InfoRecog,2,
+         "Immediate verification: found extra kernel element(s)!");
+    if FindKernelFastNormalClosure(ri,5,5) = fail then
+        ErrorNoReturn("Very bad: image was wrongly recognised ",
+                      "and  we found out too late");
+    fi;
+    Info(InfoRecog,2,"Have now ",Length(gensN(ri)),
+         " generators for kernel.");
+    return false;
+  end );
+
+InstallGlobalFunction( FindKernelRandom,
+  function(ri,n)
+    Info(InfoRecog,2,"Creating ",n," random generators for kernel.");
+    return GenerateRandomKernelElementsAndOptionallyVerifyThem(ri, n, false);
   end );
 
 InstallGlobalFunction( FindKernelDoNothing,
   function(ri,n1,n2)
     return true;
+  end );
+
+InstallGlobalFunction(RECOG_HandleSpecialCaseKernelTrivialAndMarkedForImmediateVerification,
+  function(ri)
+    # Due to the design decision that trivial kernels are not represented via
+    # recognition nodes with size one but via the value `fail`, we need to
+    # employ a bit of a hack.
+    # Since the responsible method findgensNmeth(ri) only found trivial
+    # generators the list gensN(ri) is empty. Thus KernelRecogNode(ri) would be
+    # set to fail.
+    # However, since immediateverification(ri) is set to true, we need
+    # to double-check now, whether the kernel indeed is trivial. Since
+    # KernelRecogNode(ri) is not bound yet, we can not use the function
+    # ImmediateVerification. Instead, we call FindKernelFastNormalClosure.
+    Info(InfoRecog, 2, "Found only trivial generators for the kernel but the ",
+         "kernel is marked for immediate verification. Hence we call ",
+         "FindKernelFastNormalClosure.");
+    if fail = FindKernelFastNormalClosure(
+        ri,
+        RECOG_NrElementsInImmediateVerification,
+        RECOG_FindKernelFastNormalClosureStandardArgumentValues[2]
+    ) then
+        ErrorNoReturn("Very bad: image was wrongly recognised ",
+                      "and  we found out too late");
+    fi;
   end );
 
 # Returns the product of a subsequence of a list (of generators).
@@ -125,8 +200,8 @@ end );
 InstallGlobalFunction( FindKernelFastNormalClosure,
   # Used in the generic recursive routine.
   function(ri,n1,n2)
-    if not FindKernelRandom(ri, n1) then
-        return false;
+    if FindKernelRandom(ri, n1) = fail then
+        return fail;
     fi;
 
     SetgensN(ri,FastNormalClosure(ri!.gensHmem,gensN(ri),n2));