… tst/broken/quick/bugfix.tst
… tst/working/slow/bugfix.tst