Enables immediate verification and fixes the following issues:
Fixes #11, fixes #16, fixes #35, fixes #37.

To make the immediate verification work a few fixes to other functions are needed. The main commit is f45c7e0a1ca2bf8abd.

Previously this PR also did some other things. I've put these other things into separate PRs such that this one now only contains the commits related to immediate verification.

If we merge this, I think we won't need mandarins anymore. If we want to have the "generate a hundred random elements and send them through the recognition tree" we could do the following: we could add a "thorough verification" option which does what immediate verification does for the root of a recog tree, but with more random elements. If that fails maybe restart the whole computation and if that also doesn't do the trick, tell the user.