@ssiccha 2021-06-29
Ok, so I added setting immediate verification in the `GoProjective` method.  I understand it such that immediate verification is a simpler (and less powerful) form of mandarins. You generate random elements of the kernel and check whether you can compute an SLP for those. If you can't you add them to the kernel.

This somewhat fixes the first example in #37:

    repeat ri:=RecognizeGroup(GL(9,5));; until Size(ri) <> Size(GL(9,5));

I ran this on my machine thrice until it stopped. Without setting immediate verification in `GoProjective`, on my machine it failed every time after roughly 100 runs.  I tried it once with only setting immediate verification in the `GoProjective` method, and not increasing the number of random elements immediate verification uses. Then recog was able to do roughly 1400 successfull recognitions.

With setting the number of random elements immediate verification uses to 10, I have three processes which ran 3000 successfull recognitions and which are still at it. I'll have to check whether this slows down the test-suite by a lot, but that would surprise me.