@ssiccha 2021-06-29
This also fixes #16 in that I could do 1500 and 7000 successfull recognitions once I added immediate verification to the method `NonTransitive`. In addition to doing that, there it'd probably also make sense to add a special "compute kernel generators" function which does sifting or something. Or maybe do that instead of immediate verification?