@ssiccha 2021-06-29
It might also make sense to always do an immediate verification when the kernel generators are created randomly.