@ssiccha 2021-06-29
Another thought: we could add a "thorough verification" option which does what immediate verification does for the root of a recog tree, just with 100 random elements. If that fails maybe restart the whole computation and if that also doesn't do the trick, tell the user.

I wonder in which situations mandarins would provide a significant improvement compared to immediate verification. I'm sure there are some, but for now immediate verification seems to be good enough.