@ssiccha 2021-07-01
I also got a fix for #11, but I need to think about that a little more. I'm pushing my WIP, that needs to be cleaned up though.