@ssiccha 2021-07-01
I've converted this into a draft again since I also rewrote the `FindKernelRandom` and the immediate verification code. I'll clean this up and add tests for the issues I claim this fixes. Then this should be good I think.