@fingolfin 2021-08-23
I do not understand yet how this removes the the need for mandarins, but we can discuss this via Jitsi on Wednesday