@fingolfin 2021-09-06
I'd like to know how this affects performance. Do the tests get noticeably slower / faster? Even if they are slower, we may want to merge it, but whatever happens should be tested and documented here.

As I explained on a conference call 1-2 weeks ago, I am unconvinced this is a proper replacement for Mandarins -- I am in particular concerned that getting to a confidence level similar to what e.g. 100 mandarins provide, one would have to create soooo many random elements with this method (growing rapidly with the number of branch points) that it would get very slow. But I might be wrong! But that's why I am also asking about performance