@ssiccha 2021-09-06
Hrmpf, I cleaned up the handling of the case where the kernel was recognized as being trivial, but now `veryslow/MatFDPM.tst` fails. I'll have to check whether that's by accident or whether it's systematically broken.