@ssiccha 2021-10-05
> I'd like to know how this affects performance. Do the tests get noticeably slower / faster? Even if they are slower, we may want to merge it, but whatever happens should be tested and documented here.

With this PR `tst/working/quick/bugfix.tst` takes roughly 63 seconds, instead of 200ms without this PR. Thus I'm moving that test into the `slow/` folder.

General performance doesn't seem to be affected too much by this PR. 
Here are the timings with (PR) and without (master) immediate verification. The full outputs of the test runs can be found below.

  | With PR | Master
-- | -- | --
All | 496.89 | 434.78
Quick | 19.64 | 18.3


## Test runs
### With PR
#### All 
```
testing: /home/sergio/projects/pkg/recog/tst/working/combined/ClassicalNaturalNam
   58308 ms (1480 ms GC) and 3.98GB allocated for combined/ClassicalNaturalNaming.tst
testing: /home/sergio/projects/pkg/recog/tst/working/combined/Sporadic.tst
   24806 ms (5439 ms GC) and 2.75GB allocated for combined/Sporadic.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/GenericSnAnUnknownDegree.tst
      22 ms (0 ms GC) and 4.73MB allocated for quick/GenericSnAnUnknownDegree.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatDiagonal.tst
    3178 ms (745 ms GC) and 580MB allocated for quick/MatDiagonal.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatReducible.tst
     169 ms (5 ms GC) and 14.6MB allocated for quick/MatReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatSn.tst
     472 ms (24 ms GC) and 74.7MB allocated for quick/MatSn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatTrivial.tst
       6 ms (0 ms GC) and 544KB allocated for quick/MatTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/Methods.tst
       1 ms (0 ms GC) and 27.8KB allocated for quick/Methods.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermAllSubgroups.tst
     614 ms (112 ms GC) and 52.1MB allocated for quick/PermAllSubgroups.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermCycle.tst
       2 ms (0 ms GC) and 215KB allocated for quick/PermCycle.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermDirProd.tst
      15 ms (0 ms GC) and 2.12MB allocated for quick/PermDirProd.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermThrowAwayFixedPoints.tst
       4 ms (0 ms GC) and 651KB allocated for quick/PermThrowAwayFixedPoints.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermTrivial.tst
       2 ms (0 ms GC) and 161KB allocated for quick/PermTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex.tst
     292 ms (88 ms GC) and 22.2MB allocated for quick/ProjLowIndex.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex2.tst
    2920 ms (63 ms GC) and 200MB allocated for quick/ProjLowIndex2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjNotAbsIrred.tst
     260 ms (8 ms GC) and 26.1MB allocated for quick/ProjNotAbsIrred.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjReducible.tst
     241 ms (11 ms GC) and 24.5MB allocated for quick/ProjReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjStabChain.tst
    1387 ms (44 ms GC) and 107MB allocated for quick/ProjStabChain.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield.tst
     396 ms (94 ms GC) and 28.9MB allocated for quick/ProjSubfield.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield2.tst
     379 ms (12 ms GC) and 34.1MB allocated for quick/ProjSubfield2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield3.tst
      68 ms (0 ms GC) and 4.78MB allocated for quick/ProjSubfield3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensor.tst
    1084 ms (36 ms GC) and 91.0MB allocated for quick/ProjTensor.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensorInduced.tst
    1029 ms (36 ms GC) and 85.8MB allocated for quick/ProjTensorInduced.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjThreeLargeElOrders.tst
     738 ms (25 ms GC) and 47.2MB allocated for quick/ProjThreeLargeElOrders.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTrivial.tst
       5 ms (0 ms GC) and 541KB allocated for quick/ProjTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/RecogMethod.tst
       0 ms (0 ms GC) and 59.8KB allocated for quick/RecogMethod.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/bugfix.tst
   62147 ms (6203 ms GC) and 7.26GB allocated for quick/bugfix.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/cgo1.tst
    1107 ms (19 ms GC) and 66.1MB allocated for quick/cgo1.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/classical.tst
     947 ms (20 ms GC) and 57.0MB allocated for quick/classical.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/mixed.tst
       1 ms (0 ms GC) and 54.3KB allocated for quick/mixed.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/sl3.tst
      97 ms (0 ms GC) and 9.78MB allocated for quick/sl3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/C3C5.tst
    6068 ms (217 ms GC) and 385MB allocated for slow/C3C5.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/GenericSnAnUnknownDegree.tst
   74913 ms (11461 ms GC) and 7.09GB allocated for slow/GenericSnAnUnknownDegree.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatAn.tst
   15675 ms (263 ms GC) and 1010MB allocated for slow/MatAn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatC3.tst
     295 ms (9 ms GC) and 27.0MB allocated for slow/MatC3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatC3_2.tst
    9532 ms (192 ms GC) and 771MB allocated for slow/MatC3_2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatHSmax5.tst
    2798 ms (81 ms GC) and 370MB allocated for slow/MatHSmax5.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/NonDeletedPermModuleRepOfSn.tst
    2418 ms (44 ms GC) and 130MB allocated for slow/NonDeletedPermModuleRepOfSn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/PermLargeBasePrimitive.tst
    4748 ms (1574 ms GC) and 2.63GB allocated for slow/PermLargeBasePrimitive.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/ProjC6.tst
    6590 ms (246 ms GC) and 421MB allocated for slow/ProjC6.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/ProjDet.tst
   16073 ms (316 ms GC) and 604MB allocated for slow/ProjDet.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/basic.tst
    6247 ms (1442 ms GC) and 1.04GB allocated for slow/basic.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/ClassicalNatural.ts
   82053 ms (8368 ms GC) and 7.31GB allocated for veryslow/ClassicalNatural.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/MatC6.tst
   28641 ms (1385 ms GC) and 1.92GB allocated for veryslow/MatC6.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/MatFDPM.tst
   30307 ms (646 ms GC) and 1.23GB allocated for veryslow/MatFDPM.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/PermSym10Subgroup.t
   41292 ms (3291 ms GC) and 4.67GB allocated for veryslow/PermSym10Subgroup.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/PermTrans16.tst
    8549 ms (760 ms GC) and 962MB allocated for veryslow/PermTrans16.tst
-----------------------------------
total    496896 ms (44759 ms GC) and 46.0GB allocated
              0 failures in 47 files
```

#### Quick
```
testing: /home/sergio/projects/pkg/recog/tst/working/combined/ClassicalNaturalN\
aming.tst
    2845 ms (61 ms GC) and 179MB allocated for ClassicalNaturalNaming.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/GenericSnAnUnknownDe\
gree.tst
      45 ms (0 ms GC) and 8.35MB allocated for GenericSnAnUnknownDegree.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatDiagonal.tst
    2985 ms (555 ms GC) and 535MB allocated for MatDiagonal.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatReducible.tst
     157 ms (5 ms GC) and 14.5MB allocated for MatReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatSn.tst
     447 ms (22 ms GC) and 70.7MB allocated for MatSn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatTrivial.tst
       6 ms (0 ms GC) and 541KB allocated for MatTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/Methods.tst
       0 ms (0 ms GC) and 27.8KB allocated for Methods.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermAllSubgroups.tst
     606 ms (98 ms GC) and 52.1MB allocated for PermAllSubgroups.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermCycle.tst
       2 ms (0 ms GC) and 218KB allocated for PermCycle.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermDirProd.tst
      18 ms (0 ms GC) and 2.52MB allocated for PermDirProd.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermThrowAwayFixedPo\
ints.tst
       5 ms (0 ms GC) and 653KB allocated for PermThrowAwayFixedPoints.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermTrivial.tst
       8 ms (6 ms GC) and 175KB allocated for PermTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex.tst
     244 ms (8 ms GC) and 29.2MB allocated for ProjLowIndex.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex2.tst
    3940 ms (301 ms GC) and 318MB allocated for ProjLowIndex2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjNotAbsIrred.tst
     294 ms (14 ms GC) and 31.3MB allocated for ProjNotAbsIrred.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjReducible.tst
     358 ms (98 ms GC) and 28.0MB allocated for ProjReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjStabChain.tst
    1530 ms (37 ms GC) and 117MB allocated for ProjStabChain.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield.tst
     361 ms (14 ms GC) and 36.7MB allocated for ProjSubfield.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield2.tst
     464 ms (97 ms GC) and 35.6MB allocated for ProjSubfield2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield3.tst
      65 ms (0 ms GC) and 4.71MB allocated for ProjSubfield3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensor.tst
    1030 ms (33 ms GC) and 88.3MB allocated for ProjTensor.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensorInduced.ts\
t
    1023 ms (34 ms GC) and 84.8MB allocated for ProjTensorInduced.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjThreeLargeElOrde\
rs.tst
     755 ms (18 ms GC) and 47.2MB allocated for ProjThreeLargeElOrders.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTrivial.tst
       5 ms (0 ms GC) and 540KB allocated for ProjTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/RecogMethod.tst
       0 ms (0 ms GC) and 59.8KB allocated for RecogMethod.tst
testing: /home/sergio/projects/pkg/recog/tst/working/combined/Sporadic.tst
     463 ms (245 ms GC) and 74.7MB allocated for Sporadic.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/cgo1.tst
     920 ms (14 ms GC) and 53.0MB allocated for cgo1.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/classical.tst
     989 ms (17 ms GC) and 57.0MB allocated for classical.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/mixed.tst
       0 ms (0 ms GC) and 55.0KB allocated for mixed.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/sl3.tst
      82 ms (4 ms GC) and 11.4MB allocated for sl3.tst
-----------------------------------
total     19647 ms (1681 ms GC) and 1.83GB allocated
              0 failures in 30 files
```

### Master
#### All
```
testing: /home/sergio/projects/pkg/recog/tst/working/combined/ClassicalNaturalN\
aming.tst
   62698 ms (1403 ms GC) and 3.98GB allocated for combined/ClassicalNaturalNami\
ng.tst
testing: /home/sergio/projects/pkg/recog/tst/working/combined/Sporadic.tst
   25945 ms (5336 ms GC) and 2.75GB allocated for combined/Sporadic.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/GenericSnAnUnknownDe\
gree.tst
      25 ms (0 ms GC) and 4.73MB allocated for quick/GenericSnAnUnknownDegree.t\
st
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatDiagonal.tst
    3708 ms (1402 ms GC) and 643MB allocated for quick/MatDiagonal.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatReducible.tst
     117 ms (5 ms GC) and 12.2MB allocated for quick/MatReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatSn.tst
     368 ms (13 ms GC) and 47.0MB allocated for quick/MatSn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatTrivial.tst
       6 ms (0 ms GC) and 541KB allocated for quick/MatTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/Methods.tst
       0 ms (0 ms GC) and 27.8KB allocated for quick/Methods.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermAllSubgroups.tst
     725 ms (198 ms GC) and 52.1MB allocated for quick/PermAllSubgroups.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermCycle.tst
       2 ms (0 ms GC) and 217KB allocated for quick/PermCycle.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermDirProd.tst
      17 ms (0 ms GC) and 2.14MB allocated for quick/PermDirProd.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermThrowAwayFixedPo\
ints.tst
       5 ms (0 ms GC) and 634KB allocated for quick/PermThrowAwayFixedPoints.ts\
t
testing: /home/sergio/projects/pkg/recog/tst/working/quick/PermTrivial.tst
       2 ms (0 ms GC) and 179KB allocated for quick/PermTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex.tst
     182 ms (9 ms GC) and 19.5MB allocated for quick/ProjLowIndex.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjLowIndex2.tst
    3137 ms (71 ms GC) and 216MB allocated for quick/ProjLowIndex2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjNotAbsIrred.tst
     284 ms (13 ms GC) and 28.4MB allocated for quick/ProjNotAbsIrred.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjReducible.tst
     228 ms (13 ms GC) and 23.4MB allocated for quick/ProjReducible.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjStabChain.tst
    1832 ms (145 ms GC) and 113MB allocated for quick/ProjStabChain.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield.tst
     331 ms (11 ms GC) and 29.7MB allocated for quick/ProjSubfield.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield2.tst
     411 ms (12 ms GC) and 32.2MB allocated for quick/ProjSubfield2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjSubfield3.tst
      66 ms (2 ms GC) and 4.56MB allocated for quick/ProjSubfield3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensor.tst
    1032 ms (126 ms GC) and 73.3MB allocated for quick/ProjTensor.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTensorInduced.ts\
t
    1054 ms (31 ms GC) and 84.9MB allocated for quick/ProjTensorInduced.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjThreeLargeElOrde\
rs.tst
     757 ms (15 ms GC) and 46.0MB allocated for quick/ProjThreeLargeElOrders.ts\
t
testing: /home/sergio/projects/pkg/recog/tst/working/quick/ProjTrivial.tst
       4 ms (0 ms GC) and 538KB allocated for quick/ProjTrivial.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/RecogMethod.tst
       0 ms (0 ms GC) and 59.8KB allocated for quick/RecogMethod.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/bugfix.tst
     184 ms (86 ms GC) and 10.1MB allocated for quick/bugfix.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/cgo1.tst
    1039 ms (19 ms GC) and 57.6MB allocated for quick/cgo1.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/classical.tst
    1047 ms (18 ms GC) and 57.0MB allocated for quick/classical.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/mixed.tst
       1 ms (0 ms GC) and 54.3KB allocated for quick/mixed.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/sl3.tst
      76 ms (3 ms GC) and 11.8MB allocated for quick/sl3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/C3C5.tst
    6397 ms (221 ms GC) and 380MB allocated for slow/C3C5.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/GenericSnAnUnknownDeg\
ree.tst
   85302 ms (12172 ms GC) and 7.50GB allocated for slow/GenericSnAnUnknownDegre\
e.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatAn.tst
    6916 ms (136 ms GC) and 569MB allocated for slow/MatAn.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatC3.tst
     314 ms (7 ms GC) and 25.9MB allocated for slow/MatC3.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatC3_2.tst
    8284 ms (158 ms GC) and 525MB allocated for slow/MatC3_2.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/MatHSmax5.tst
    2729 ms (277 ms GC) and 309MB allocated for slow/MatHSmax5.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/NonDeletedPermModuleR\
epOfSn.tst
   10423 ms (420 ms GC) and 576MB allocated for slow/NonDeletedPermModuleRepOfS\
n.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/PermLargeBasePrimitiv\
e.tst
    4178 ms (1548 ms GC) and 2.29GB allocated for slow/PermLargeBasePrimitive.t\
st
testing: /home/sergio/projects/pkg/recog/tst/working/slow/ProjC6.tst
    8013 ms (297 ms GC) and 479MB allocated for slow/ProjC6.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/ProjDet.tst
   18053 ms (344 ms GC) and 602MB allocated for slow/ProjDet.tst
testing: /home/sergio/projects/pkg/recog/tst/working/slow/basic.tst
    6371 ms (1505 ms GC) and 1.01GB allocated for slow/basic.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/ClassicalNatural.\
tst
   73646 ms (7048 ms GC) and 5.96GB allocated for veryslow/ClassicalNatural.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/MatC6.tst
   43961 ms (2115 ms GC) and 2.64GB allocated for veryslow/MatC6.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/MatFDPM.tst
    2616 ms (274 ms GC) and 144MB allocated for veryslow/MatFDPM.tst
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/PermSym10Subgroup\
.tst
   43694 ms (3277 ms GC) and 4.67GB allocated for veryslow/PermSym10Subgroup.ts\
t
testing: /home/sergio/projects/pkg/recog/tst/working/veryslow/PermTrans16.tst
    8608 ms (772 ms GC) and 954MB allocated for veryslow/PermTrans16.tst
-----------------------------------
total    434788 ms (39502 ms GC) and 36.8GB allocated
              0 failures in 47 files
```

#### Quick
```
testing: /home/sergio/projects/pkg/recog/tst/working/combined/ClassicalNaturalNaming.tst
    2917 ms (63 ms GC) and 179MB allocated for ClassicalNaturalNaming.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/GenericSnAnUnknownDegree.tst
      48 ms (0 ms GC) and 8.35MB allocated for GenericSnAnUnknownDegree.tst
testing: /home/sergio/projects/pkg/recog/tst/working/quick/MatDiagonal.tst
    2767 ms (607 ms GC) and 532MB allocated for MatDiagonal.tst
…